refactor(Piece): Piece를 인터페이스로 변경, EmptyPiece와 NoneEmptyPiece로 구분
