+++ src/main/java/chess/domain/movement/Mobility.java
@@ -2,6 +2,5 @@ package chess.domain.movement;
 
 public enum Mobility {
     ONCE,
-    INFINITE,
-    EMPTY
+    INFINITE
 }
