+++ src/main/java/chess/domain/movement/Movement.java
@@ -11,8 +11,7 @@ public enum Movement {
     ROOK(Mobility.INFINITE, Direction.getCross()),
     BISHOP(Mobility.INFINITE, Direction.getDiagonal()),
     KNIGHT(Mobility.ONCE, Direction.getLShaped()),
-    PAWN(Mobility.ONCE, Direction.getNorthern()),
-    EMPTY(Mobility.EMPTY, Direction.getEmpty());
+    PAWN(Mobility.ONCE, Direction.getNorthern());
 
     private final Mobility mobility;
     private final List<Direction> directions;
