+++ src/main/java/chess/domain/piece/King.java
@@ -3,7 +3,7 @@ package chess.domain.piece;
 import chess.domain.movement.Movement;
 import chess.domain.Team;
 
-public class King extends Piece {
+public class King extends NoneEmptyPiece {
 
     public King(final Team team) {
         super(team, Movement.KING);
