+++ src/main/java/chess/domain/piece/Knight.java
@@ -3,7 +3,7 @@ package chess.domain.piece;
 import chess.domain.movement.Movement;
 import chess.domain.Team;
 
-public class Knight extends Piece {
+public class Knight extends NoneEmptyPiece {
 
     public Knight(final Team team) {
         super(team, Movement.KNIGHT);
