+++ src/main/java/chess/domain/piece/Queen.java
@@ -3,7 +3,7 @@ package chess.domain.piece;
 import chess.domain.movement.Movement;
 import chess.domain.Team;
 
-public class Queen extends Piece {
+public class Queen extends NoneEmptyPiece {
 
     public Queen(final Team team) {
         super(team, Movement.QUEEN);
