+++ src/main/java/chess/domain/piece/Rook.java
@@ -3,7 +3,7 @@ package chess.domain.piece;
 import chess.domain.movement.Movement;
 import chess.domain.Team;
 
-public class Rook extends Piece {
+public class Rook extends NoneEmptyPiece {
 
     public Rook(final Team team) {
         super(team, Movement.ROOK);
